pyproject.toml
[project]
name = "genai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio-client>=1.3.0",
    "huggingface-hub>=0.36.0",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.1",
    "streamlit-drawable-canvas>=0.9.3",
    "torch>=2.9.1",
    "transformers>=4.46.3",
]
